Moved SolidFoundations.Framework.Integrations.DGAIntegration to SolidFoundations.Framework.External.SpaceCore
